backend/requirements.txt
#Python version : 3.12.10
asgiref==3.9.1
Django==5.0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.5
djangorestframework==3.14.0
django-taggit==5.0.1
mysqlclient==2.2.1
python-dotenv==1.0.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2